pyproject.toml
[tool.poetry]
name = "ranks-bot"
version = "0.1.0"
description = "A rank bot for discord.gg/diabetes"
authors = ["ImVaskel <49348256+ImVaskel@users.noreply.github.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
psutil = "^5.8.0"
discord-ext-menus = {git = "https://github.com/rapptz/discord-ext-menus"}
discord-ext-menus-views = {git = "https://github.com/oliver-ni/discord-ext-menus-views"}
"discord.py" = {git = "https://github.com/rapptz/discord.py.git", extras = ["speedups"]}
PyYAML = "^6.0"
Pillow = "^9.1.0"
jishaku = "^2.5.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"
pyright = "^1.1.354"
pre-commit = "^3.6.2"
poethepoet = "^0.25.0"
asyncpg-stubs = "^0.29.1"


[tool.poetry.group.docker.dependencies]
docker = "^7.0.0"

[tool.poe.tasks]
all = ["format", "lint", "check"]
check = "pyright"
format = "ruff format ."
lint = "ruff check ."

[tool.ruff]
target-version = "py311"
line-length = 125

[tool.ruff.lint]
select = [
    "C4",
    "F",
    "G",
    "I",
    "PTH",
    "RUF",
    "SIM",
    "TCH",
    "UP",
    "W",
    "PERF",
    "ANN",
]
ignore = [
    "F401",
    "F402",
    "F403",
    "F405",
    "PERF203",
    "PLR2004",
    "RUF001",
    "RUF009",
    "SIM105",
    "S101",
    "UP034",
    "UP038",
    "ANN101",
    "ANN102",
    "ANN401",
]

[tool.ruff.lint.isort]
split-on-trailing-comma = true
combine-as-imports = true

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.pyright]
useLibraryCodeForTypes = true
typeCheckingMode = "basic"
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
ignore = []
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
strictParameterNoneValue = true
reportMissingImports = "error"
reportUnusedImport = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"
reportUnusedExpression = "error"
reportGeneralTypeIssues = "error"
reportDuplicateImport = "error"
reportUntypedFunctionDecorator = "error"
reportUnnecessaryTypeIgnoreComment = "warning"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
